requirements.txt
transformers
langchain
together
pyttsx3              
SpeechRecognition    
python-dotenv
requests
pyaudio         
